requirements.txt
flask==3.0.3
flask-socketio==5.3.6
Flask-Cors==4.0.1
eventlet==0.36.1
nltk==3.8.1
transformers==4.44.0
torch==2.4.0
numpy==1.24.1
requests==2.32.3 
# sqlalchemy==2.0.32 
# pymysql==1.1.1 
# tidb_vector==0.0.10
python-dotenv==1.0.1
groq==0.9.0
pydub
faster-whisper==1.0.3
langchain-groq==0.1.9
sentence-transformers
nvidia-cublas-cu11 
nvidia-cudnn-cu11
pymongo[srv]